// src/types.ts
export interface SentryTeam {
  id: string;
  name: string;
  slug: string;
}

export interface SentryProject {
  id: string;
  name: string;
  slug: string;
  platform?: string;
  teams: SentryTeam[];
  environments?: string[];
  features?: string[];
}
// ...
export interface SentryProjectIssue {
  annotations: any[];
  assignedTo: {
    type: string;
    id: string;
    name: string;
    email: string;
  } | null;
  count: string;
  culprit: string;
  firstSeen: string;
  hasSeen: boolean;
  id: string;
  isBookmarked: boolean;
  isPublic: boolean;
  isSubscribed: boolean;
  lastSeen: string;
  level: string;
  logger: string | null;
  metadata: {
    title: string;
  };
  numComments: number;
  permalink: string;
  project: {
    id: string;
    name: string;
    slug: string;
  };
  shareId: string | null;
  shortId: string;
  stats: {
    "24h": [number, number][];
  };
  status: string;
  statusDetails: Record<string, any>;
  subscriptionDetails: any | null;
  title: string;
  type: string;
  userCount: number;
  isUnhandled?: boolean;
  platform?: string;
}
// ...
export interface SentryIssueActivity {
  data: Record<string, any>;
  dateCreated: string;
  id: string;
  type: string;
  user: {
    id: string;
    name: string;
    email: string;
  } | null;
}

export interface SentryReleaseProject {
  name: string;
  slug: string;
}

export interface SentryRelease {
  authors: any[];
  commitCount: number;
  data: Record<string, any>;
  dateCreated: string;
  dateReleased: string | null;
  deployCount: number;
  firstEvent: string;
  lastCommit: any | null;
  lastDeploy: any | null;
  lastEvent: string;
  newGroups: number;
  owner: any | null;
  projects: SentryReleaseProject[];
  ref: string | null;
  shortVersion: string;
  url: string | null;
  version: string;
}

export interface SentryIssueDetailsResponse {
  activity: SentryIssueActivity[];
  annotations: string[];
  assignedTo: {
    id: string;
    name: string;
    email: string;
  } | null;
  count: string;
  culprit: string;
  firstRelease: SentryRelease | null;
  firstSeen: string;
  hasSeen: boolean;
  id: string;
  isBookmarked: boolean;
  isPublic: boolean;
  isSubscribed: boolean;
  lastRelease: SentryRelease | null;
  lastSeen: string;
  level: string;
  logger: string | null;
  metadata: {
    title: string;
  };
  numComments: number;
  participants: any[];
  permalink: string;
  pluginActions: any[];
  pluginContexts: any[];
  pluginIssues: any[];
  project: {
    id: string;
    name: string;
    slug: string;
  };
  seenBy: any[];
  shareId: string | null;
  shortId: string;
  stats: {
    "24h": [number, number][];
    "30d": [number, number][];
  };
  status: string;
  statusDetails: Record<string, any>;
  subscriptionDetails: any | null;
  tags: {
    key: string;
    value: string;
  }[];
  title: string;
  type: string;
  userCount: number;
  userReportCount: number;
}